(feat) added East Coast time conversion which handles Daylight Saving - added gates for delete account and edit account info

+++ app/Actions/Jetstream/DeleteUser.php
@@ -4,6 +4,7 @@ namespace App\Actions\Jetstream;
 
 use App\Models\User;
 use Laravel\Jetstream\Contracts\DeletesUsers;
+use Illuminate\Support\Facades\Gate;
 
 class DeleteUser implements DeletesUsers
 {
@@ -12,6 +13,10 @@ class DeleteUser implements DeletesUsers
      */
     public function delete(User $user): void
     {
+        if (Gate::denies('make-changes')) {
+            session()->flash('error', 'Demo users cannot make changes.');
+            return;
+        }
         $user->deleteProfilePhoto();
         $user->tokens->each->delete();
         $user->delete();
